Move problem model classes to problem package
