+++ src/main/java/com/rocketden/main/dao/ProblemRepository.java
@@ -2,7 +2,7 @@ package com.rocketden.main.dao;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.rocketden.main.model.Problem;
+import com.rocketden.main.model.problem.Problem;
 
 // This will be AUTO IMPLEMENTED by Spring into a Bean called problemRepository
 // CRUD refers Create, Read, Update, Delete
