+++ src/main/java/com/rocketden/main/dto/problem/ProblemMapper.java
@@ -1,7 +1,7 @@
 package com.rocketden.main.dto.problem;
 
-import com.rocketden.main.model.Problem;
-import com.rocketden.main.model.ProblemTestCase;
+import com.rocketden.main.model.problem.Problem;
+import com.rocketden.main.model.problem.ProblemTestCase;
 import org.modelmapper.ModelMapper;
 
 public class ProblemMapper {
