+++ src/main/java/com/rocketden/main/dto/room/UpdateSettingsRequest.java
@@ -1,7 +1,7 @@
 package com.rocketden.main.dto.room;
 
 import com.rocketden.main.dto.user.UserDto;
-import com.rocketden.main.model.ProblemDifficulty;
+import com.rocketden.main.model.problem.ProblemDifficulty;
 import lombok.Getter;
 import lombok.Setter;
