+++ src/main/java/com/rocketden/main/model/problem/ProblemTestCase.java
@@ -1,4 +1,4 @@
-package com.rocketden.main.model;
+package com.rocketden.main.model.problem;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
